Change main to SPROLARunner to meet the requirement
